% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patch_data.R
\name{get_patch_data}
\alias{get_patch_data}
\title{Add patch characteristics to initial data frame created by med_to_dt function}
\usage{
get_patch_data(
  dat,
  leverOmission = F,
  removeOmission = T,
  removeIncomplete = T
)
}
\arguments{
\item{dat}{data table; created by med_to_dt function}

\item{leverOmission}{logical; were there lever timeouts, default=false}

\item{removeOmission}{logical; remove omissions from data table; default=true}

\item{removeIncomplete}{logical; remove incomplete patches from data table; default=true}
}
\value{
Create patch data frame w/ variables: \cr
 PatchNum = patch number visited during session \cr
 PressPatch = presses since last reset \cr
 PressPatch_Inv = number of presses until reset \cr
 StartVolume = patch starting volume \cr
 PatchTime = patch time \cr
 PatchReward = patch reward \cr
 PatchRate = patch rate \cr
 MVT_AvRate = average reward rate according to MVT \cr
}
\description{
Add patch characteristics to initial data frame created by med_to_dt function
}
