% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forage_data.R
\name{load_forage_data}
\alias{load_forage_data}
\title{Load data frame of foraging data that was previously saved using save_forage_data}
\usage{
load_forage_data(file)
}
\arguments{
\item{file}{string or vector of strings; complete path to file name or vector of file paths. If a vector, returns one data frame.}
}
\description{
Load data frame of foraging data that was previously saved using save_forage_data
}
