% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_to_dt.R
\name{med_to_dt}
\alias{med_to_dt}
\title{Convert MedPC File to an R data table}
\usage{
med_to_dt(
  file,
  travel = 10,
  group = "",
  treatment = "",
  varITI = F,
  ITI = 7,
  opto = F
)
}
\arguments{
\item{file}{string; complete path to MedPC data file}

\item{travel}{numeric; travel time for experiment}

\item{group}{string; experimental group of animal}

\item{treatment}{string; experimental treatment (i.e. drug)}

\item{varITI}{logical; if true, ITI = ITI-decision time-handling time-reward time. If false, ITI = ITI}

\item{ITI}{numeric; ITI for session}

\item{opto}{logical, was optogenetics used?}
}
\value{
data table with the following variables: \cr
 Subject \cr
 Group \cr
 Date = ddmmyy \cr
 Travel = travel time used for session \cr
 Treatment \cr
 Trial = trial number \cr
 Decision = 0 if lever press, 1 if nosepoke \cr
 DecisionRT = reaction time to make decision \cr
 DecisionTime = time stamp of decision during session \cr
 HandlingTime = time between lever press and reward port entry; 0 if nosepoke \cr
 RewardTime = time stamp on reward port entry \cr
 RewardVolume_s = reward volume in s of pump time \cr
 RewardVolume_mL = reward volume in mL \cr
 Omission = if removeOmission = F, 0 if no reward omission, 1 if omission \cr
 ITI = intertrial interval; 7 if lever press, 0 if nosepoke \cr
 TravelTime = travel time on trial; 0 if lever press, 'Travel' if nosepoke \cr
 TrialTime = total time for trial = DecisionRT+HandlingTime+ITI+TravelTime \cr
 CumulativeTime = total time foraging/day \cr
 CumulativeReward_mL = cumulative reward intake/day \cr
 InstantRate = RewardVolume_mL/TrialTime \cr
 CumulativeRate = cumulativeTime/CumulativeReward_mL \cr
}
\description{
Convert MedPC File to an R data table
}
